Use short display for ship headers in !fleet
Captain review: ship lines should also be emoji-only short format,
matching the bot lines. Full fleet output is now compact tree.

+++ src/relay.ts
@@ -3615,7 +3615,7 @@ function registerRelayCommands(): void {
               commissioned,
               health: worstGrade,
               tokPerDay: shipTok,
-            }, 'long'));
+            }, 'short'));
           }
 
           // Warn when a commissioned ship has active bots but no health report — relay likely offline (#135)
